+++ src/main/java/duke/Ui.java
@@ -1,3 +1,6 @@
+package duke;
+import duke.exception.DukeException;
+
 import java.util.Scanner;
 
 public class Ui {
