+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,9 +1,16 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.InvalidArgumentException;
+
 public class DeleteCommand extends Command{
     public DeleteCommand(String command) {
         super(command);
     }
     @Override
-    public void execute(TaskList tasks ,Ui ui, Storage storage) {
+    public void execute(TaskList tasks , Ui ui, Storage storage) {
         String[] words = this.fullCommand.split(" ", 2);
         try {
             tasks.deleteTask(Integer.parseInt(words[1]) - 1);
