+++ src/main/java/duke/command/ExitCommand.java
@@ -1,10 +1,17 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.DukeException;
+
 import java.io.IOException;
 public class ExitCommand extends Command {
     public ExitCommand(String command) {
         super(command);
     }
     @Override
-    public void execute(TaskList tasks ,Ui ui, Storage storage) {
+    public void execute(TaskList tasks , Ui ui, Storage storage) {
         Ui.showLine();
         System.out.println("Bye. Hope to see you again soon!");
         Ui.showLine();
