+++ src/main/java/duke/exception/InvalidDateException.java
@@ -1,3 +1,4 @@
+package duke.exception;
 public class InvalidDateException extends DukeException {
     public InvalidDateException() {
         super("☹ OOPS!!! I'm sorry, but your date is in the incorrect format! " +
