+++ src/main/java/duke/exception/InvalidTaskException.java
@@ -1,3 +1,4 @@
+package duke.exception;
 public class InvalidTaskException extends DukeException {
     public InvalidTaskException() {
         super("☹ OOPS!!! I'm sorry, but I don't know what that means :-(");
